Refactor getNodeRoles into @console/shared
