+++ frontend/packages/console-shared/src/selectors/index.ts
@@ -1,4 +1,5 @@
 export * from './common';
-export * from './machine';
 export * from './infrastructure';
 export * from './pod';
+export * from './machine';
+export * from './node';
