+++ frontend/public/module/k8s/node.ts
@@ -28,18 +28,3 @@ const isNodeReady = (node: K8sResourceKind): boolean => {
 };
 
 export const nodeStatus = (node: K8sResourceKind) => (isNodeReady(node) ? 'Ready' : 'Not Ready');
-
-const NODE_ROLE_PREFIX = 'node-role.kubernetes.io/';
-export const getNodeRoles = (node: K8sResourceKind): string[] => {
-  const labels = _.get(node, 'metadata.labels');
-  return _.reduce(
-    labels,
-    (acc: string[], v: string, k: string) => {
-      if (k.startsWith(NODE_ROLE_PREFIX)) {
-        acc.push(k.slice(NODE_ROLE_PREFIX.length));
-      }
-      return acc;
-    },
-    [],
-  );
-};
